+++ packages/PIPS/pips/src/Libs/phrase/comEngine_generate_HRECode.c
@@ -35,11 +35,13 @@ one process on the HRE
 #include "genC.h"
 #include "linear.h"
 #include "ri.h"
+#include "effects.h"
 
 #include "resources.h"
 
 #include "misc.h"
 #include "ri-util.h"
+#include "effects-util.h"
 #include "pipsdbm.h"
 
 #include "text-util.h"
@@ -97,9 +99,9 @@ expression get_fifo_from_ref(reference ref)
 {
   expression retExp = expression_undefined;
 
-  int fifoNum = (int)hash_get(gRefToHREFifo, ref);
+  intptr_t fifoNum = (intptr_t)hash_get(gRefToHREFifo, ref);
 
-  if(fifoNum != (int)HASH_UNDEFINED_VALUE)
+  if(fifoNum != (intptr_t)HASH_UNDEFINED_VALUE)
     {
       retExp = make_integer_constant_expression(fifoNum);
     }
