+++ packages/PIPS/pips/src/Libs/prettyprint/cprettyprinter.c
@@ -47,12 +47,14 @@
 #include "genC.h"
 #include "linear.h"
 #include "ri.h"
+#include "effects.h"
 
 #include "resources.h"
 #include "properties.h"
 
 #include "misc.h"
 #include "ri-util.h"
+#include "effects-util.h"
 #include "pipsdbm.h"
 #include "text-util.h"
 #include "transformations.h"
@@ -314,8 +316,8 @@ static string c_dim_string(list ldim)
         {
             expression elow = dimension_lower(dim);
             expression eup = dimension_upper(dim);
-            int low;
-            int up;
+            intptr_t low;
+            intptr_t up;
             string slow;
             string sup;
 
@@ -929,7 +931,7 @@ static string c_reference(reference r)
       expression e_tmp;
       expression e_lower = dimension_lower(DIMENSION(CAR(l_dim)));
       string s;
-      int itmp;
+      intptr_t itmp;
 
       if( !expression_equal_integer_p(e_lower, 0)) 
 	e_tmp =
